Hooks folder added for creating custom hooks useSaveFile hook added for saving the current file
