+++ public/src/components/WindowBarControlsSaveLogic.js
@@ -1,22 +1,16 @@
 import React from 'react'
-import {useDispatch , useSelector} from 'react-redux'
-import postContent from '../thunks/postFileContent'
 import WindowBarControlsSave from './WindowBarControlsSave'
-import hashFileContent from '../helpers/hashFileContent'
-import { current } from '@reduxjs/toolkit'
+import useSaveFile from '../hooks/useSaveFile'
 
 export default function WindowBarControlsSaveLogic(props){
-    const dispatch  = useDispatch()
-    const currentFile = useSelector(state => state.files.currentFile)
-    const workspaceId = useSelector(state => state.workspace.workspaceId)
+    const saveFile = useSaveFile()
+
     function saveClickHandler(e){
-        const hash = hashFileContent(currentFile.content)
-        if(hash !== currentFile.savedHash){
-            e.preventDefault()
-            dispatch(postContent({nanoId : currentFile.id , content :  currentFile.content , workspaceId }))
-        }
+        e.preventDefault()
+        saveFile()
     }
     return (
         <WindowBarControlsSave saveClickHandler={saveClickHandler} />
     )
 }
+
